# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Stemmarest REST API",
    "description" : "Stemmarest is a Neo4j-based repository for variant text traditions (i.e. texts that have been transmitted in multiple manuscripts). For an introduction to its conceptual structure, read on; to dive into specific API calls, see the navigation menu, which is organised by the sort of object each call concerns.\n\n### Definition of terms\n\nWhat follows is a definition of the concepts that occur within Stemmarest; understanding these terms will be helpful for understanding the API documentation (each is also described again, in the same words, at the top of its own section in the navigation menu). Note that most of these entities are defined with respect to (therefore, under) the **/tradition/{tradId}** heading in the API; readings, although defined as root objects, always belong to one and only one tradition section.\n",
    "version" : "1.1"
  },
...
  "paths" : {
...
    "/tradition/{tradId}/annotationlabel/{name}" : {
      "get" : {
        "tags" : [ "Annotation Label" ],
        "summary" : "Get annotation label spec",
        "description" : "Retrieves the specification for the given annotation type name.",
        "operationId" : "getAnnotationLabel",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tradId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnnotationLabelModel"
                }
              }
            }
          },
          "400" : {
            "description" : "if the annotation label reference is a name shared by multiple labels (legacy data only)",
            "content" : {
              "application/json" : { }
            }
          },
          "404" : {
            "description" : "Annotation label not found",
            "content" : {
              "application/json" : { }
            }
          },
          "500" : {
            "description" : "Failure, with an error report in JSON format",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Annotation Label" ],
        "summary" : "Put annotation label spec",
        "description" : "Creates or updates an annotation type specification.",
        "operationId" : "createOrUpdateAnnotationLabel",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tradId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "The AnnotationLabelModel specification to use",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AnnotationLabelModel"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated existing label",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnnotationLabelModel"
                }
              }
            }
          },
          "201" : {
            "description" : "Created new label",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnnotationLabelModel"
                }
              }
            }
          },
          "400" : {
            "description" : "if the annotation label reference is a name shared by multiple labels (legacy data only)",
            "content" : {
              "application/json" : { }
            }
          },
          "409" : {
            "description" : "Requested name is already in use, or a rename was requested for a label that is still in use by an annotation or another label's links",
            "content" : {
              "application/json" : { }
            }
          },
          "500" : {
            "description" : "Failure, with an error report in JSON format",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Annotation Label" ],
        "summary" : "Delete annotation label",
        "description" : "Deletes the specified annotation label specification from the tradition. Returns an error if the label is still in use.",
        "operationId" : "deleteAnnotationLabel",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tradId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnnotationLabelModel"
                }
              }
            }
          },
          "400" : {
            "description" : "if the annotation label reference is a name shared by multiple labels (legacy data only)",
            "content" : {
              "application/json" : { }
            }
          },
          "404" : {
            "description" : "Annotation label not found",
            "content" : {
              "application/json" : { }
            }
          },
          "409" : {
            "description" : "Annotation label is still in use",
            "content" : {
              "application/json" : { }
            }
          },
          "500" : {
            "description" : "Failure, with an error report in JSON format",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AnnotationLabelModel" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "properties" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "links" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
...
    }
  }
}